requirements.txt
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-dotenv==1.0.0
requests==2.31.0
pandas==2.1.4
openpyxl==3.1.2
PyYAML==6.0.1
PyPDF2==3.0.1
jsonschema==4.20.0
gunicorn==21.2.0
psycopg2-binary==2.9.9
whitenoise==6.6.0
celery==5.3.4
redis==5.0.1
dj-database-url==2.1.0
